requirements.txt
flask==3.0.3
flask-login==0.6.3
ultralytics==8.3.29
fpdf2==2.7.4
opencv-python==4.10.0.84
pillow==10.4.0
pytz==2024.1